Rearrange directories
