+++ shared/src/androidMain/kotlin/org/hyperskill/app/project_selection/injection/PlatformProjectSelectionListComponentImpl.kt
@@ -1,6 +1,7 @@
 package org.hyperskill.app.project_selection.injection
 
 import org.hyperskill.app.core.injection.ReduxViewModelFactory
+import org.hyperskill.app.project_selection.list.injection.ProjectSelectionListComponent
 import org.hyperskill.app.project_selection.presentation.ProjectSelectionListViewModel
 import ru.nobird.app.presentation.redux.container.wrapWithViewContainer
