+++ shared/src/commonMain/kotlin/org/hyperskill/app/project_selection/details/injection/ProjectSelectionDetailsParams.kt
@@ -1,0 +1,10 @@
+package org.hyperskill.app.project_selection.details.injection
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ProjectSelectionDetailsParams(
+    val trackId: Long,
+    val projectId: Long,
+    val isProjectSelected: Boolean
+)
