--- shared/src/commonMain/kotlin/org/hyperskill/app/project_selection/domain/analytic/ProjectsSelectionListViewedHyperskillAnalyticEvent.kt
@@ -1,22 +1,0 @@
-package org.hyperskill.app.project_selection.domain.analytic
-
-import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
-import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
-import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
-
-/**
- * Represents a view analytic event for the projects selection list screen.
- *
- * JSON payload:
- * ```
- * {
- *     "route": "/tracks/1/projects",
- *     "action": "view"
- * }
- * ```
- * @see HyperskillAnalyticEvent
- */
-class ProjectsSelectionListViewedHyperskillAnalyticEvent(trackId: Long) : HyperskillAnalyticEvent(
-    HyperskillAnalyticRoute.Tracks.Projects(trackId),
-    HyperskillAnalyticAction.VIEW
-)
