--- shared/src/commonMain/kotlin/org/hyperskill/app/project_selection_details/injection/ProjectSelectionDetailsParams.kt
@@ -1,10 +1,0 @@
-package org.hyperskill.app.project_selection_details.injection
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ProjectSelectionDetailsParams(
-    val trackId: Long,
-    val projectId: Long,
-    val isProjectSelected: Boolean
-)
