+++ shared/src/commonTest/kotlin/org/hyperskill/projects_selection/ProjectsListTest.kt
@@ -9,19 +9,19 @@ import org.hyperskill.ResourceProviderStub
 import org.hyperskill.app.core.view.mapper.NumbersFormatter
 import org.hyperskill.app.core.view.mapper.SharedDateFormatter
 import org.hyperskill.app.profile.domain.model.Profile
-import org.hyperskill.app.project_selection.domain.analytic.ProjectsSelectionListClickedProjectHyperskillAnalyticsEvent
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature.Action
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature.ContentState
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature.InternalAction
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature.Message
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature.State
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListFeature.ViewState
-import org.hyperskill.app.project_selection.presentation.ProjectSelectionListReducer
-import org.hyperskill.app.project_selection.presentation.bestRatedProjectId
-import org.hyperskill.app.project_selection.presentation.fastestToCompleteProjectId
-import org.hyperskill.app.project_selection.presentation.recommendedProjects
-import org.hyperskill.app.project_selection.view.mapper.ProjectSelectionListViewStateMapper
+import org.hyperskill.app.project_selection.list.domain.analytic.ProjectsSelectionListClickedProjectHyperskillAnalyticsEvent
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature.Action
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature.ContentState
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature.InternalAction
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature.Message
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature.State
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListFeature.ViewState
+import org.hyperskill.app.project_selection.list.presentation.ProjectSelectionListReducer
+import org.hyperskill.app.project_selection.list.presentation.bestRatedProjectId
+import org.hyperskill.app.project_selection.list.presentation.fastestToCompleteProjectId
+import org.hyperskill.app.project_selection.list.presentation.recommendedProjects
+import org.hyperskill.app.project_selection.list.view.mapper.ProjectSelectionListViewStateMapper
 import org.hyperskill.app.projects.domain.model.Project
 import org.hyperskill.app.projects.domain.model.ProjectKind
 import org.hyperskill.app.projects.domain.model.ProjectLevel
